Support local pictures and better chromatic aberration correction

+++ app/src/main/assets/syros.frag
@@ -9,16 +9,16 @@ uniform int LensLimits;
 
 //uniform sampler2D Texture0;
 
-varying mediump vec4 oColor;
-varying mediump vec2 oTexCoord0;
-varying mediump vec2 oTexCoord1;
-varying mediump vec2 oTexCoord2;
+varying highp vec4 oColor;
+varying highp vec2 oTexCoord0;
+varying highp vec2 oTexCoord1;
+varying highp vec2 oTexCoord2;
 
 void main()
 {
-    mediump float ResultA = 0.0;;
+    highp float ResultA = 0.0;;
 
-    mediump float ResultR;
+    highp float ResultR;
 
     if (oTexCoord0.x > 1.0 || oTexCoord0.y > 1.0 || oTexCoord0.x < 0.0 || oTexCoord0.y < 0.0)
     {
@@ -31,7 +31,7 @@ void main()
         ResultA = texture2D(Texture0, oTexCoord1).a;
     }
 
-    mediump float ResultG;
+    highp float ResultG;
     if (oTexCoord1.x > 1.0 || oTexCoord1.y > 1.0 || oTexCoord1.x < 0.0 || oTexCoord1.y < 0.0)
     {
          ResultG = (LensLimits == 1 ? 0.2 : 0.0);
@@ -41,7 +41,7 @@ void main()
         ResultG = texture2D(Texture0, oTexCoord1).g;
     }
 
-    mediump float ResultB;
+    highp float ResultB;
     if (oTexCoord2.x > 1.0 || oTexCoord2.y > 1.0 || oTexCoord2.x < 0.0 || oTexCoord2.y < 0.0)
     {
          ResultB = (LensLimits == 1 ? 0.3 : 0.0);
@@ -51,6 +51,5 @@ void main()
         ResultB = texture2D(Texture0, oTexCoord2).b;
     }
 
-    gl_FragColor = vec4(ResultR *  oColor.r, ResultG * oColor.g , ResultB * oColor.b, ResultA);
-    //gl_FragColor = vec4((ResultR + 0.1) *  oColor.r, (ResultG  + 0.2) * oColor.g, (ResultB + 0.3) * oColor.b , 1.0);
+    gl_FragColor = vec4(ResultR *  oColor.r, ResultG * oColor.g, ResultB * oColor.b, ResultA);
 }
